poprawki zgodnie z uwagami na wideo YT
