+++ README.md
@@ -2,3 +2,6 @@
 Rozwiązania do zadania z Newslettera #2 - 6 marca 2019
 
 Do Newslettera można zapisać się na stronie: https://mailtrain.javastart.pl/subscription/tc5pDEzUq
+
+
+https://youtu.be/ZuZGAl4bDAA?t=1713 - omówienie mojego rozwiązania
